Handle Error parsing the same way as Response
